requirements/_base.txt
Django==3.2
psycopg2==2.9.3
Pillow==9.2.0
django-imagekit==4.1.0
